fix: support formatIndent
